test(api): add unit tests for medicine verification route
