+++ apps/api/package.json
@@ -7,7 +7,7 @@
     "start": "node dist/index.js",
     "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
     "build": "tsc",
-    "test": "tsc -p tsconfig.test.json && node --test \".tmp/test-dist/**/*.test.js\""
+    "test": "jest"
   },
   "dependencies": {
     "@supabase/supabase-js": "^2.105.4",
@@ -25,9 +25,14 @@
   "devDependencies": {
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
+    "@types/jest": "^30.0.0",
     "@types/morgan": "^1.9.9",
     "@types/node": "^25.7.0",
+    "@types/supertest": "^7.2.0",
     "@types/web-push": "^3.6.4",
+    "jest": "^30.4.2",
+    "supertest": "^7.2.2",
+    "ts-jest": "^29.4.9",
     "ts-node-dev": "^2.0.0",
     "typescript": "^6.0.3"
   }
