+++ apps/api/tests/setup.ts
@@ -1,0 +1,2 @@
+process.env.SUPABASE_URL = "http://localhost:54321";
+process.env.SUPABASE_ANON_KEY = "test-anon-key";
